requirements.txt
django
django-restframework
django-cors-headers